Move RedirectStdOutAndErr
