+++ subprojects/logging/src/test/groovy/org/gradle/internal/logging/slf4j/OutputEventListenerBackedLoggerDefaultConfigurationTest.groovy
@@ -18,7 +18,7 @@ package org.gradle.internal.logging.slf4j
 
 import org.gradle.api.logging.Logging
 import org.gradle.internal.time.Time
-import org.gradle.util.RedirectStdOutAndErr
+import org.gradle.util.internal.RedirectStdOutAndErr
 import org.gradle.util.TextUtil
 import org.junit.Rule
 import org.slf4j.Logger
