+++ subprojects/tooling-api/src/crossVersionTest/groovy/org/gradle/integtests/tooling/r23/StandardStreamsCrossVersionSpec.groovy
@@ -22,7 +22,7 @@ import org.gradle.integtests.tooling.fixture.ToolingApiLoggingSpecification
 import org.gradle.integtests.tooling.fixture.ToolingApiVersion
 import org.gradle.tooling.ProjectConnection
 import org.gradle.util.GradleVersion
-import org.gradle.util.RedirectStdOutAndErr
+import org.gradle.util.internal.RedirectStdOutAndErr
 import org.junit.Rule
 
 class StandardStreamsCrossVersionSpec extends ToolingApiLoggingSpecification {
